+++ internal/subscriber/models.go
@@ -1,11 +1,15 @@
 package subscriber
 
-import "time"
+import (
+	"time"
+
+	"gorm.io/gorm"
+)
 
 type Subscriber struct {
 	ID         string `gorm:"primary_key"`
 	CreatedAt  time.Time
 	UpdatedAt  time.Time
-	DeletedAt  time.Time
+	DeletedAt  gorm.DeletedAt `gorm:"index"`
 	WebhookURL string
 }